updated the controller form for users

+++ app/controllers/users_controller.rb
@@ -2,6 +2,8 @@
 
 class UsersController < ApplicationController
   before_action :authenticate_user!
+  after_action :verify_authorized
+
   def index
     @users = User.all
     authorize User
@@ -14,7 +16,28 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
+    authorize user
     user.destroy
     redirect_to users_path, notice: 'Nutzer Gelöscht'
   end
+
+  def update
+    @user = User.find(params[:id])
+    authorize @user
+
+      if @user.update_attributes(secure_params)
+        flash[:success] = "Benutzer erfolgreich aktualisiert"
+        redirect_to @user
+      else
+        flash[:error] = "nicht in der Lage zu aktualisieren"
+        render "user"
+      end
+  end
+
+  private
+
+  def secure_params
+    params.require(:user).permit(:role)
+  end
+  
 end
